// src/app/generated_services/model/showGraduationRequirementsDTO.ts
/**
 * Carlson Gracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: support@carlsongracie.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShowBeltDTO } from './showBeltDTO';


export interface ShowGraduationRequirementsDTO { 
    id?: string;
    belt?: ShowBeltDTO | null;
    beltId: string;
    description: string;
    minimumClasses?: number;
}